Added Some Helper Paragraphs For Tracking
There is a problem with the configuration, so I have added some helper paragraphs to let me know where the content is showing up.  Thi sis a temporary addition as I can't debug this at the moment but I want to be able to pick this up and easily see what is happening.