Basic Setup For Index and Posts
This is the basic setup for future tinkering. Everything mostly works as expected at this point, but I still need to create a working Image component to add to the BlogPost component.  Currently just grabbing the first picture.  If there are multiple pictures, this setup won't work.  Also, I would like to add some prev / next buttons for posts.