More Design Changes
I'm still trying to figure out the fonts which aren't being imported properly.  Additionally, I'm having trouble with some of the positioning and layout on the index page.

Next moves are to add an About page so that I can fix that link on the index.  I also need to add a preview field in contentful and use that in the Featured Post.  I also want to try fixing the images as they aren't displaying in an aesthetically pleasing way.  Possibly needs a component?  Styled components are a bit tricky for this stuff.  Finally, I want to revisit the font issues as that will help the look and feel of the page.

Remove Strange Font Import Setup

I tried to import fonts on a few ways that didn't work.  After consulting the official Gatsby docs, I discovered that they recommend a global.css file and importing fonts from GoogleFonts directly.  Once I got that working, removing the other methods was quick and easy.

Refactor Blog Post Into Various Implementations

While it appeared in the beginning that there was only one use case for the BlogPost component, it became clear as I continued to build and brainstorm that there were actually three use cases.  So, I have created a component for the following cases: featured on the index page, full post on its own page, and simplified for the All Posts page.

Now I am realizing that there is a lot of code that has been duplicated.  So it's time to create some additional components to streamline updates.  I'll be starting with the following: Date, Tags, Title, Image, Paragraph, TruncatedParagraph, and Links.  May need more later... LOL

Reduce Duplication With Components

This commit implements the changes prescribed in the last commit.  There is an issue with tiling the pictures for the individual blog post, but hopefully I can figure that out.

The only other page that needs imaginging is the About This Blog page.  Not quite sure what I want that page to look like.  I also need to add a horizontal rule between the posts on the All Posts page.